Split test_views
